Moving the payment request exception to the exceptions directory
